Refactored and updated tests
